Update rate limit for enrollment api
LEARNER-5101: The current rate limit was adjusted to be
very high for staff users to account for excessive calls
from the ecommerce service. The ecommerce service has
been updated to not make these calls. This commit intends
to adjust the rate limit to it's original state before the
excessive calls.