Move update_discussions_map to django_comment_common
The task to update the mapping of inline discussion XBlocks to
discussion_ids was originally implemented in an LMS-only process, and
did not properly catch a course_published signal from Studio. I tried
to enable the Discussions app in Studio (it's a pluggable Django app),
but this resulted in difficult to debug bok choy errors around cohorts.

This commit moves the task into django_comment_common, which is also
where the model it's altering is located. It is possible to just move
the listener to django_comment_common and keep the task in LMS and
bridge the two with routing, but I thought that would be unnecessarily
confusing.