Allow scrolling after insert/edit image in Studio
The scroll lock class `modal-open` is removed from the body when the
EditImageModal is closed using the "Close" or "X" button, but not when the
"Insert Image" button is clicked. The submit button handler now calls the
`closeImageModal` handler too so that the `modal-open` class is removed after
inserting or editing an image.