[edxnotes] use the correct url and update the API params
This is really two separate things:

1. Fix the notes retirement endpoint to generate the correct notes API
url "/api/v1/annotations/" rather than the nonexistent "/api/v1/".

2. Update the API params to use "user" rather than "user_id".  This
depends on another PR in edx/edx-notes-api to make the corresponding
change on the API side.  This is required because "user" is the
hard-coded key that the permissions class uses to check the JWT token
user.