avoid fetching large json fields on site_configuration filtering
On Tahoe we have ~800 SiteConfiguration objects. The `sass_variables`
and `page_elements` fields are very large and cause the queries here
to take almost 2 seconds, plus another two seconds of JSON decoding
everytime they are called.

This should fix a good chunk of the performance problems we are seeing
on Tahoe.