Let notify_credentials get args from db
To allow the notify_credentials management command to run from a
jenkins job, we want to allow the command to pull its arguments
from the database, rather than the jenkins interface.

So this adds a table that admins can adjust in the DB and a new
argument for notify_credentials to tell it to go look there.

LEARNER-6196
LEARNER-6197