Add Contact Us form customization options
This commit adds personalization options for the Contact Us links on 
edx-platform's
page.
To allow for this customization, two SiteConfiguration variables were 
added:
* CONTACT_US_ENABLE: Switch to enable/disable the Contact Us page. 
Setting this to
False will disable the contact page and the links on the footer will 
disappear.

* CONTACT_US_CUSTOM_LINK: If the contact page is enabled, this setting 
allows to set
a custom URL for the Contact Us links on edx-platform. If this setting 
is not set,
the platform will use the default contact form.