Skip tests of help doc links on named release branches.
@skip_unless_master is used to skip tests because on named release
branches, most work happens leading up to the first release on the
branch, and that is before the docs have been published.  Tests that
check readthedocs for the right doc page will fail during this time, and
it's just a big distraction.  Also, if we bork the docs, it's not the
end of the world, and we can fix it easily, so this is a good tradeoff.