ENT-1054: fix enrollment page when for non-standard course run ids
The enterprise enrollment page is failing for course runs that have
non-standard ids.  For the sake of this PR, 'non-standard ids' refers to
cases where the ```course_id``` is not a substring of ```course_run_id```.
For example, if there is a  course ```edx+DemoX``` that has a course run
```edx+DemoY+run1```, the enrollment page would not populate the right
data for this course run.  One of the reasons it's failing is because
it calls the ```enrollment``` API to get a list of available modes and
the ```enrollment``` API does not handle non-standard ```course_run_id```s
gracefully.

```edx-platform``` assumes that ```course_id```s are always substrings
of ```course_run_id```s but this is not always the case.  Using the
above example, if you call the ```enrollment``` API with course run
```edx+DemoY+run1``` you will not get the correct results.
The ```enrollment``` API needs to get data associated with the course and
it will try to parse the ```course_id``` from the ```course_run_id```.
Since ```edx+DemoY``` is not a course the ```enrollment``` API can't find
the right course and can't get the right data.  To fix this we to get
the canonical ```course_id``` by calling the ```course_run``` API in
discovery.

I created a method ```get_course_locator(course_run_id)``` that creates
a ```CourseLocator``` object based on the given ```course_run_id```.
This is just like ```CourseKey.from_string``` but with one important
distinction.  The ```org``` and ```course``` fields are populated based
on the ```course_id``` NOT the ```course_run_id```.  So in our example
```get_course_locator``` would create a ```CourseLocator``` object with
```org=edx```, ```course=DemoX``` and ```run=run1``` while
```CourseKey.from_string``` would have created a ```CourseLocator``` object
with ```org=edx```, ```course=DemoY``` and ```run=run1```.

Remaining work:
 - This does not yet work with the old course id style
 - Writing tests
 - Arch review