An initial breakdown of our areas of the monolith.
Right now there are no owners associated but this should clarify what is
considered part of the core and core extensions.  This is not a full
list so things will have to be added as we find that they need an owner
and don't have one.  This list was taken by looking at the parts of our
codebase that is most executed for edx.org and so should represent the
most important things that should probably have owners.