Add location ID to Studio component footers
This is helpful for a number of use cases for us at Stanford,
where instructors/staff/admin need access to an item's location id.

Previously, we'd have to dig this value out by other means.
This makes it easy to discover and use as needed, eg: copy/paste, etc.

One major example of our use of it is with the `in-video-quiz` XBlock [1][2],
which maps video components to capa problems.

- [1] https://github.com/Stanford-Online/xblock-in-video-quiz
- [2] https://github.com/Stanford-Online/xblock-in-video-quiz/blob/master/invideoquiz/invideoquiz.py#L41-L61