Import lms.djangoapps.courseware, not courseware.
In the early days, we did some hacky path munging to make separate
import trees for LMS vs. CMS (a.k.a. Studio). Some of this lingers
to today. We want to get rid of that hacky path munging, so all
imports of the courseware app should use the real full path of
"lms.djangoapps.courseware", not just "courseware".

This commit does that conversion for edx-platform. I didn't remove
the deprecation warning because it's likely that there are other
repos that are bundled in edx-platform that do use the old form.