Reference XBlock mixins by name.
Don't import them in settings files because they try to load django
settings as a part of their loading. This causes django to throw an
error reasonably saying it's SECRET_KEY is not set, which is true
because we haven't finished loading the settings yet.

Making the change in just the CMS first because the LMS change is
causing some failures so will do that separately.