Continue to pass the InheritanceMixin class object.
When we pass it in as a string, it seems to be causing a bunch of test
failures.

When I run `paver test_system -s lms -t lms/djangoapps/courseware/tests/test_module_render.py::TestTOC`
I get failures in the tests for split.  I'm not exactly sure what's
going on yet but it looks like there is a default value for a field that
is overridden by the course factory.  But that override doesn't seem to
be applied to the split course?

Testing with this now to see if there are any other issues.