Rmove the extra configurabilty logic for reset buttons.
The `DEFAULT_SHOW_RESET_BUTTON` is not actually referenced anywhere
other than in these two places where it's read.  As far as I can tell,
nothing sets this value so it's always going to fallback to false.

Since this mixin is imported during startup, removing this unblocks
testing on python3 where it can fail to start up if the settings haven't
fully loaded before you try to access them.