DEPR-15: Remove usage of firebase-token-generator
The old Notes Django app is the only caller of this no longer
supported library. I have verified that this old app is no
longer functional and is already slated to be removed as part
of DEPR-18.

So this commit simply removes the call to firebase-token-generator
and leaves removal of the Notes app to DEPR-18.