Convert to using ExitStack instead of nested. (#21329)
* Convert to using ExitStack instead of nested.

nested is not available in Python3.

* Use ExitStack from contextlib2.

ExitStack has not been backported to the Python 2.7 standard library and
contextlib.nested is no longer available in Python 3.5.  We need to be
able to use something in both 2 and 3 as we transition so we'll pull in
contextlib2 which has backports of the full python3 contextlib.