Set a default value for PROCTORING_SETTINGS
The proctoring app assumes that this setting exists and so we get an
attribute error at runtime if it doesn't.  In python 2 this was not an
issue but because of the change in how exceptions are handle for getting
attributes, it's a problem in python 3.

The correct thing to do would be to fix this in the proctoring app so
that it checks for the existence of its value before using it but that's
a longer cycle to make/deploy that change so doing this as a stopgap to
get past it.