Remove the mongo plugin for the djnago-debug-toolbar.
We are currently running this plugin from a has directly off of master
to get support for a newer version of DJDT.  However even on master
there is not yet support for python 3.  We're running into this when
running runserver and when collecting static assets.

A quick survey of the development team at edX shows that people are no
longer actively using this debug panel.(14 responders all saying no to
the question "In the last year have you used the mongo portion of the
django-debug-toolbar in edx-platform?")

This makes sense since the module store and courseware related
developement is not very active right now.  Given all this, I'm removing
this toolbar for now and if we need it again we can survey what's
available at that time to choose the best tool.