Skip tests that can't succeed in Python3 + Django 1.11
There is a bug in the underlying management/base code that tries to make
all manageent command output be unicode.  This management command
outputs the binary tar file data and so breaks in python3.  In python2
the code is happy to pass bytes back and forth and in later versions of
django this is fixed.  Howevere it's not possible to get this test to
pass in Python3 and django 1.11