Use the default json session serializer.
This will force a logout as sessions fail to load but this should be a
more performant and secure serializer moving forward.  The reason we
overwrote it in our config is that it previously used to be the default
and we didn't want things to breake and force logouts when we changed
it.  We're no more okay with people getting logged out.