Hide Studio's re-index button for non-global staff
This commit was originally submitted in PR #19610 by OmarIthawi.
It hides the button that manually triggers courseware search re-
indexing unless you're a global staff user. The rationale and
discussion for this can be found at:

https://github.com/edx/edx-platform/pull/19610

But the short version is that it was:

A) Dangerous if it was working correctly.
B) It would 403 if you tried it without global staff access.