Remove this test because it is flaky.
It failed and passed on the same commit.  The failure was that the page
timed out, so it could be fixed by making that page performant.  I'd
rather not increase the timeout for the page to load as the bokchoy
tests are already very slow.

I'm also not sure why this test needs to exist.  I would think that this
could be tested on the underlying function without testing at the UI
layer.